Move Create... UI functions to creation.cpp
Moved all of the general UI creating functions to a separate file.
These functions are not changed frequently and just clutter up the actually relevant code.
